+++ tests/tests_main.c
@@ -31,8 +31,8 @@ int main(int argc, char** argv)
     // registerTest("Monitor Mode Test", monitorModeTest);
     // registerTest("Window Test", windowTest);
     // registerTest("Icon Test", iconTest);
-    registerTest("Cursor Test", cursorTest);
-    // registerTest("Input Window Test", inputWindowTest);
+    // registerTest("Cursor Test", cursorTest);
+    registerTest("Input Window Test", inputWindowTest);
     // registerTest("System Cursor Test", systemCursorTest);
     // registerTest("Attach Window Test", attachWindowTest);
     // registerTest("Character Event Test", charEventTest);