+++ CHANGELOG.md
@@ -76,6 +76,9 @@ reflecting its role as the primary explicit foundation for OS and graphics abstr
 - **Video:** Added Wayland-based backend support.
 - **Video:** Added **palGetVideoFeaturesEx()** to check old and extended supported features.
 - **Video:** Added **palGetWindowHandleInfoEx()** to get extended window handles.
+- **Video:** Added **palGetRawMouseWheelDelta()** to get raw mouse wheel delta.
+- **Core:** Added **palPackFloat()** to combine two floats into a single Int64 integer.
+- **Core:** Added **palUnpackFloat()** to retreive two floats from a single Int64 integer.
 
 ### Tests
 - Added native integration example: demonstrating **Native API Integration with PAL API**. see **native_integration_test.c**.